+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support for setting path to service key file in config
 * Ability to translate into multiple comma-separated languages
 * Ability to set source language with `from:<code>`
+* Hiding of result when source is set to auto and translation matches the
+  detected language, as long as there are at least two target languages
 
 ### Changed
 * Invalid language codes are no longer left for the API to catch